backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
ultralytics
opencv-python
numpy
torch
